+++ src/demo-app/app/layout/layout.menu.ts
@@ -26,6 +26,7 @@ export const OPTIONS: MenuOption[] = [
    { route: 'help-demo', name: 'help' },
    { route: 'info-box-demo', name: 'info-box' },
    { route: 'item-list-demo', name: 'item-list' },
+   { route: 'form-demo', name: 'Form' },
    { route: 'main', name: 'main' },
    { route: 'modal-demo', name: 'Modal' },
    { route: 'pagination-demo', name: 'Pagination' },