+++ src/egeo-demo/public_api.ts
@@ -15,6 +15,7 @@
  */
 
 export { StAlertsDemoModule } from './st-alert-demo/st-alerts-demo.module';
+export { StFormDemoModule } from './st-form-demo/st-form-demo.module';
 export { StTwoListSelectionDemoModule } from './st-two-list-selection-demo/st-two-list-selection-demo.module';
 export { StTooltipDemoModule } from './st-tooltip-demo/st-tooltip-demo.module';
 export { StSwitchDemoModule } from './st-switch-demo/st-switch-demo.module';