+++ src/lib/barrels.ts
@@ -23,6 +23,7 @@ import { StDropdownModule } from './st-dropdown/st-dropdown.module';
 import { StFooterModule } from './st-footer/st-footer.module';
 import { StFormDirectiveModule } from './directives/form/form-directives.module';
 import { StFormLabelModule } from './utils/egeo-form/st-form-label/st-form-label.module';
+import { StFormModule } from './st-form/st-form.module';
 import { StHeaderModule } from './st-header/st-header.module';
 import { StHelpModule } from './st-help/st-help.module';
 import { StHorizontalTabsModule } from './st-horizontal-tabs/st-horizontal-tabs.module';
@@ -61,6 +62,7 @@ export const DECLARATIONS: any[] = [
    StFooterModule,
    StFormDirectiveModule,
    StFormLabelModule,
+   StFormModule,
    StHeaderModule,
    StHelpModule,
    StHorizontalTabsModule,